# packages/shared/src/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/admin/events": {
      "get": {
        "description": "Drafts first (\"Rascunho · Data a definir\"), then chronological.",
        "operationId": "AdminEventsController_list_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminEventsResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Eventos: gradient cards with valor chip, date, inscritos/arrecadado totals",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "description": "Valor vazio = gratuito. status=published requires date + local (422 event.publish_requirements).",
        "operationId": "AdminEventsController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminEventDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create an event (Rascunho by default; publishing is a separate gesture)",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/events/{id}": {
      "patch": {
        "description": "A published event must keep its date and local (422 event.publish_requirements).",
        "operationId": "AdminEventsController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminEventDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Edit an event (canceled events are frozen history — 409)",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/events/{id}/publish": {
      "post": {
        "description": "Students never see an undated event: missing date/local → 422 event.publish_requirements.",
        "operationId": "AdminEventsController_publish_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminEventDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Publish a draft (audited) — requires date + local",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/events/{id}/cancel": {
      "post": {
        "description": "Voids the open event charges and cancels pending registrations — nobody is billed for a dead event. Confirmed rows keep their history.",
        "operationId": "AdminEventsController_cancel_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminEventDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel an event (never hard-delete; audited)",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/events/{id}/registrations": {
      "get": {
        "operationId": "AdminEventsController_registrations_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminEventRegistrationsResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Inscritos: who registered, status, who confirmed, paid amount + totals",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/events/{id}/announce": {
      "post": {
        "description": "Emits events.announcement.requested + one audit row and NOTHING else — delivery is the notifications phase. Clients show \"Comunicado enviado aos inscritos.\" on 202.",
        "operationId": "AdminEventsController_announce_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnounceResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Comunicar: queue an announcement to the inscritos (published only)",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/aluno/events/{id}": {
      "get": {
        "description": "Published events only — drafts and canceled events behave as 404.",
        "operationId": "AlunoEventsController_detail_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlunoEventDetailResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Event detail: banner, data/local/responsável, valor chip, own state",
        "tags": [
          "aluno"
        ]
      }
    },
    "/v1/aluno/events/{id}/registration": {
      "post": {
        "description": "Gratuito = confirmed on the spot. Paid = pending_payment + an event-origin charge paid via POST /aluno/wallet/charges/{chargeId}/payments (same Pix sheet, same simulate). Cancel-and-reconfirm reuses the same row — one registration per (event, student).",
        "operationId": "AlunoEventsController_register_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterEventResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Confirmar presença (free) / Pagar inscrição (paid → chargeId)",
        "tags": [
          "aluno"
        ]
      },
      "delete": {
        "description": "Canceling a pending registration voids its open charge. A paid, confirmed registration is only undone by an admin refund (409 event.registration_settled).",
        "operationId": "AlunoEventsController_cancel_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancelar participação (free or not-yet-paid)",
        "tags": [
          "aluno"
        ]
      }
    },
    "/v1/responsavel/events": {
      "get": {
        "operationId": "ResponsavelEventsController_list_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponsavelEventsResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Published events as gradient cards with one chip per dependent",
        "tags": [
          "responsavel"
        ]
      }
    },
    "/v1/responsavel/events/{id}/registrations/{studentId}": {
      "post": {
        "description": "Same free/paid semantics as the aluno flow, billed to the guardian. Paid: pay the returned chargeId via POST /responsavel/payments/charges/{chargeId}/payments.",
        "operationId": "ResponsavelEventsController_register_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "studentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterEventResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Confirm (free) / start paying (paid) for one dependent",
        "tags": [
          "responsavel"
        ]
      },
      "delete": {
        "operationId": "ResponsavelEventsController_cancel_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "studentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel one dependent's registration (free or not-yet-paid)",
        "tags": [
          "responsavel"
        ]
      }
    },
...
  },
  "info": {
    "title": "Tatame API",
    "description": "Identity (spec 001-auth), enrollment (spec 003) and attendance (spec 004) surfaces. Errors are RFC 9457 application/problem+json with stable `code`s. The live-chamada SSE stream is a documented exception to the generated REST contract: its event payloads are component schemas, but clients wire their own streaming primitive.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "EventRegistrationStateDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending_payment",
              "confirmed",
              "canceled"
            ]
          },
          "chargeId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "Open event-origin charge to pay (pending_payment only) — drives the Pix sheet"
          }
        },
        "required": [
          "id",
          "status"
        ]
      },
...
      "EventResponsibleDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "fullName": {
            "type": "string",
            "example": "Paulo Professor"
          }
        },
        "required": [
          "userId",
          "fullName"
        ]
      },
      "EventTotalsDto": {
        "type": "object",
        "properties": {
          "inscritos": {
            "type": "number",
            "description": "Non-canceled registrations (pending + confirmed)"
          },
          "confirmados": {
            "type": "number"
          },
          "arrecadadoCents": {
            "type": "number",
            "description": "Settled event money (paid charges; refunds excluded)"
          }
        },
        "required": [
          "inscritos",
          "confirmados",
          "arrecadadoCents"
        ]
      },
      "AdminEventDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "Open mat de verão"
          },
          "bannerPreset": {
            "type": "string",
            "example": "event-purple-pink",
            "description": "Design-system gradient slug"
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "startsAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Null only on drafts (\"Data a definir\")"
          },
          "date": {
            "type": "string",
            "example": "2026-08-22",
            "nullable": true,
            "description": "Tenant-local date"
          },
          "time": {
            "type": "string",
            "example": "10:00",
            "nullable": true,
            "description": "Tenant-local time"
          },
          "priceCents": {
            "type": "number",
            "nullable": true,
            "description": "Integer cents; null = gratuito"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published",
              "canceled"
            ]
          },
          "responsible": {
            "$ref": "#/components/schemas/EventResponsibleDto"
          },
          "totals": {
            "$ref": "#/components/schemas/EventTotalsDto"
          }
        },
        "required": [
          "id",
          "name",
          "bannerPreset",
          "status",
          "responsible",
          "totals"
        ]
      },
      "AdminEventsResponseDto": {
        "type": "object",
        "properties": {
          "events": {
            "description": "Drafts first, then chronological",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminEventDto"
            }
          }
        },
        "required": [
          "events"
        ]
      },
      "CreateEventDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Open mat de verão"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "bannerPreset": {
            "type": "string",
            "example": "event-purple-pink",
            "description": "Design-system gradient catalog slug (no image upload in v1)"
          },
          "location": {
            "type": "string",
            "nullable": true,
            "example": "Tatame principal"
          },
          "startsAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time",
            "description": "ISO instant; drafts may omit it (\"Data a definir\")"
          },
          "priceCents": {
            "type": "number",
            "nullable": true,
            "minimum": 1,
            "description": "Integer cents; omitted/null = gratuito (never 0)"
          },
          "responsibleUserId": {
            "type": "string",
            "format": "uuid",
            "description": "The \"Responsável: Prof. …\" line"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published"
            ],
            "description": "Default draft — creating and publishing are separate gestures"
          }
        },
        "required": [
          "name",
          "responsibleUserId"
        ]
      },
      "UpdateEventDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "bannerPreset": {
            "type": "string",
            "example": "event-purple-pink"
          },
          "location": {
            "type": "string",
            "nullable": true,
            "description": "Null clears it on drafts only (published keeps date + local)"
          },
          "startsAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "priceCents": {
            "type": "number",
            "nullable": true,
            "minimum": 1
          },
          "responsibleUserId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "AdminRegistrationStudentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fullName": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "fullName"
        ]
      },
      "AdminRegistrationRowDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "student": {
            "$ref": "#/components/schemas/AdminRegistrationStudentDto"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending_payment",
              "confirmed",
              "canceled"
            ]
          },
          "confirmedBy": {
            "description": "Who confirmed — the aluno themself or the responsável",
            "allOf": [
              {
                "$ref": "#/components/schemas/EventResponsibleDto"
              }
            ]
          },
          "paidAmountCents": {
            "type": "number",
            "nullable": true,
            "description": "Settled amount for this registration (null while unpaid / free)"
          }
        },
        "required": [
          "id",
          "student",
          "status",
          "confirmedBy"
        ]
      },
      "AdminEventRegistrationsResponseDto": {
        "type": "object",
        "properties": {
          "event": {
            "$ref": "#/components/schemas/AdminEventDto"
          },
          "registrations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminRegistrationRowDto"
            }
          },
          "totals": {
            "$ref": "#/components/schemas/EventTotalsDto"
          }
        },
        "required": [
          "event",
          "registrations",
          "totals"
        ]
      },
      "AnnounceResponseDto": {
        "type": "object",
        "properties": {
          "recipients": {
            "type": "number",
            "description": "Inscritos addressed by the queued announcement"
          }
        },
        "required": [
          "recipients"
        ]
      },
      "AlunoEventDetailResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "Open mat de verão"
          },
          "bannerPreset": {
            "type": "string",
            "example": "event-purple-pink",
            "description": "Design-system gradient slug"
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "startsAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Null only on drafts (\"Data a definir\")"
          },
          "date": {
            "type": "string",
            "example": "2026-08-22",
            "nullable": true,
            "description": "Tenant-local date"
          },
          "time": {
            "type": "string",
            "example": "10:00",
            "nullable": true,
            "description": "Tenant-local time"
          },
          "priceCents": {
            "type": "number",
            "nullable": true,
            "description": "Integer cents; null = gratuito"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "responsible": {
            "description": "The \"Responsável: Prof. …\" line",
            "allOf": [
              {
                "$ref": "#/components/schemas/EventResponsibleDto"
              }
            ]
          },
          "registration": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/EventRegistrationStateDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "name",
          "bannerPreset",
          "responsible"
        ]
      },
      "RegisterEventResponseDto": {
        "type": "object",
        "properties": {
          "registration": {
            "$ref": "#/components/schemas/EventRegistrationStateDto"
          },
          "chargeId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "Paid events: pay this charge through the existing wallet rails (Pix sheet + simulate). Null on free events."
          }
        },
        "required": [
          "registration"
        ]
      },
...
      "ResponsavelEventsResponseDto": {
        "type": "object",
        "properties": {
          "events": {
            "description": "Published upcoming, chronological",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResponsavelEventDto"
            }
          }
        },
        "required": [
          "events"
        ]
      },
...
    }
  }
}